+++ clients/include/testing_geqrf.hpp
@@ -75,7 +75,7 @@ inline hipblasStatus_t testing_geqrf_bad_arg(const Arguments& arg)
     const int          M      = 100;
     const int          N      = 101;
     const int          lda    = 102;
-    const int          A_size = N * lda;
+    const size_t       A_size = size_t(N) * lda;
     const int          K      = std::min(M, N);
 
     host_vector<T> hA(A_size);
